Add emoji to title
Originally I wanted the emoji to be the icon for the result, however it
appears Alfred will only accept paths to files to render as icons, and
can't render the strings directly. Since it's much easier to pick the
right emoji to use when you can see it visually I'm settling for adding
it into the title.